# docs/api/OPENAPI_DRAFT.yaml
openapi: 3.1.0
info:
  title: PakimonGO API Draft
  version: 0.2.0-alpha0
  description: >
    OpenAPI draft reflecting Sprint 1-22 implemented endpoints.
    Endpoints marked `x-status: planned` are future features not yet implemented.
    This is a living document — update when new endpoints are added.
  x-versions:
    supported: [v1, v2]
    default: v1
servers:
  - url: https://api.pakimongo.example/v1
    description: Placeholder production URL
  - url: http://localhost:8000/v1
    description: Local development URL
security:
  - bearerAuth: []
paths:
...
  /users/me:
    get:
      summary: Get current user profile
      operationId: getMyProfile
      x-status: implemented
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      summary: Update current user profile
      operationId: patchMyProfile
      x-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchUserProfileRequest"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/me/collection:
    get:
      summary: Get current user species collection
      operationId: getMyCollection
      x-status: implemented
      x-requirements:
        - FR-LB-005
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: context
          in: query
          schema:
            type: string
            enum: [wild, zoo, pet, unknown]
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [totalPoints, species, captureCount, lastCaptured]
            default: totalPoints
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: include_sensitive
          in: query
          schema:
            type: boolean
            default: false
          description: Include sensitive species (requires elevated permissions)
      responses:
        "200":
          description: User collection of scored species
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCollectionResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            code: NOT_FOUND
            message: "Resource not found"
            requestId: "req_abc123"
...
  schemas:
...
    UserProfile:
      type: object
      required: [userId, email, status]
      properties:
        userId:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, suspended, deleted]
        ageBand:
          type: string
          nullable: true
          enum: [adult, teen]
        homeRegion:
          type: string
          nullable: true
        trustState:
          type: string
          nullable: true
          enum: [trusted, normal, low, restricted]
        createdAt:
          type: string
          format: date-time
          nullable: true
        totalPoints:
          type: integer
          description: Sum of every real ScoreEvent point value for this user.
...
    CollectionEntry:
      type: object
      required: [species, context, totalPoints, captureCount]
      properties:
        species:
          type: string
          example: Aquila chrysaetos
        context:
          type: string
          enum: [wild, zoo, pet, unknown]
        totalPoints:
          type: integer
        captureCount:
          type: integer
        lastCaptured:
          type: string
          format: date-time
          nullable: true
...
    PatchUserProfileRequest:
      type: object
      properties:
        ageBand:
          type: string
          enum: [adult, teen]
        homeRegion:
          type: string

    # ── Pagination ────────────────────────────────────────────
    Pagination:
      type: object
      required: [limit, offset, total]
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
    PaginatedCollectionResponse:
      type: object
      required: [userId, species, pagination]
      properties:
        userId:
          type: string
        species:
          type: array
          items:
            $ref: "#/components/schemas/CollectionEntry"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    ErrorEnvelope:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          enum:
            - UNAUTHENTICATED
            - APP_CHECK_REQUIRED
            - FORBIDDEN
            - NOT_FOUND
            - VALIDATION_FAILED
            - IDEMPOTENCY_CONFLICT
            - UPLOAD_EXPIRED
            - RATE_LIMITED
            - QUARANTINED
            - PROVIDER_UNAVAILABLE
            - TIMEOUT
            - CONFLICT
            - INTERNAL
        message:
          type: string
        requestId:
          type: string